requirements.txt
# HWM-v1 Requirements
# Lightweight dependencies for Raspberry Pi

torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
